feat(repository): Enhance repositories with additional query methods for game results and participants

+++ src/main/java/com/huy/quizme_backend/dto/game/QuestionResultDTO.java
@@ -31,7 +31,7 @@ public class QuestionResultDTO {
     @AllArgsConstructor
     public static class OptionStatDTO {
         private Long optionId;
-        private Integer percentage;
+        private Double percentage;
     }
 
     /**
